job_skill_gap_analyzer/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.8.2
requests==2.32.3

openai==1.42.0
pinecone-client==5.0.1

pdfplumber==0.11.4

pandas==2.2.2
numpy==2.0.1

streamlit==1.38.0
python-multipart==0.0.9
plotly
reportlab
